Refactor menu creation amd filling combo boxes

+++ src/by/bsuir/verkpavel/courseproject/dao/entity/Rate.java
@@ -4,61 +4,60 @@ import by.bsuir.verkpavel.courseproject.dao.Entity;
 
 import com.j256.ormlite.field.DatabaseField;
 
-
 public class Rate implements Entity {
 
-	@DatabaseField(generatedId = true)
-	private int idRate;
-	@DatabaseField
-	private float depth;
-	@DatabaseField
-	private float heigth;
-	@DatabaseField
-	private float weigth;
-	@DatabaseField
-	private float width;
+    @DatabaseField(generatedId = true)
+    private int idRate;
+    @DatabaseField
+    private int depth;
+    @DatabaseField
+    private int heigth;
+    @DatabaseField
+    private int weigth;
+    @DatabaseField
+    private int width;
 
-	public Rate() {
-	}
+    public Rate() {
+    }
 
-	public int getIdRate() {
-		return this.idRate;
-	}
+    public int getIdRate() {
+        return this.idRate;
+    }
 
-	public void setIdRate(int idRate) {
-		this.idRate = idRate;
-	}
+    public void setIdRate(int idRate) {
+        this.idRate = idRate;
+    }
 
-	public float getDepth() {
-		return this.depth;
-	}
+    public int getDepth() {
+        return this.depth;
+    }
 
-	public void setDepth(float depth) {
-		this.depth = depth;
-	}
+    public void setDepth(int depth) {
+        this.depth = depth;
+    }
 
-	public float getHeigth() {
-		return this.heigth;
-	}
+    public int getHeigth() {
+        return this.heigth;
+    }
 
-	public void setHeigth(float heigth) {
-		this.heigth = heigth;
-	}
+    public void setHeigth(int heigth) {
+        this.heigth = heigth;
+    }
 
-	public float getWeigth() {
-		return this.weigth;
-	}
+    public int getWeigth() {
+        return this.weigth;
+    }
 
-	public void setWeigth(float weigth) {
-		this.weigth = weigth;
-	}
+    public void setWeigth(int weigth) {
+        this.weigth = weigth;
+    }
 
-	public float getWidth() {
-		return this.width;
-	}
+    public int getWidth() {
+        return this.width;
+    }
 
-	public void setWidth(float width) {
-		this.width = width;
-	}
+    public void setWidth(int width) {
+        this.width = width;
+    }
 
 }
